cleaned up some errors I got in WSL

+++ src/engine/src/clock.cpp
@@ -16,5 +16,5 @@ u64 Clock::calcNodesPerSecond(u64 nodes) const
     i64 et = getElapsedTime();
     if (et < 1)
         return 0;
-    return nodes / u64(et / 1000.f);
+    return (u64)((float)nodes / (float)et / 1000.f);
 }
